[REFACTOR] Add mobile check to make sure server gets the right response
If the mobile view check is active, we have to make sure the server and
WP are in sync. If server thinks it's mobile and it's not,
need to return that it's not mobile. If WP thinks it's mobile and
server thinks it's not, need to return that it's mobile.