[NEW] Implement first version of third party integration
Any third party plugins that require special attention should have a file in the
new third party directory. We will call the hook points so that it is simple
to integrate future plugins.

First implementation is for WooCommerce.
Took the code out of class-litespeed-cache.php and into its own
file in thirdparty.