[UPDATE] Change the way advanced-cache.php checking works.
On activation, if the advanced-cache.php file does not belong to us,
we will now activate anyway.

On the admin pages, if the file does not belong to us, output a warning.
On front end pages, if the file does not belong to us, do not cache the page.

There is a new config option to skip checking for advanced-cache.php

If the file is non-existent or is blank and writable, always make it ours.