[BUGFIX] Fix possible opcode cache problem related to setting WP_CACHE
It is possible that by the time activation finishes and the redirected
page needs to be reloaded, that the opcode cache doesn't update the WP_CACHE
define. In this instance, an error message will display stating that
it was unable to write to wp-config.php, when that is not truly the case.
By adding an extra check to the error message, we should be able to avoid this.