Added DocBlock to JS Defer Exclusion Filter
Found excessive `Array to string conversion` errors coming from a custom filter that thought the filter parameter was string.  The errors stopped when the filter parameter was treated as an array.  The DocBlock makes it clear that the `litespeed_optm_js_defer_exc` filter needs an array and not a string.