Fix for PHP 7.2 Backtrace Warning
PHP 7.2 will not recognize a parameter an array parameter unless it is wrapped in the `is_array` function or is typecasted as an array.  Because of this, an excessive backtrace warning is thrown.  Adding the `(array)` typecast for the `$haystack` parameter resolves this issue. Ref: [Issue #130](https://github.com/litespeedtech/lscache_wp/issues/130)