[FEATURE] Initial implementation of Auto CDN Setup.
Auto CDN Setup is intended to simplify the process of setting up
QUIC.cloud CDN. This is a three step procedure:
1. Create/login to a QC account.
2. Click the Run Setup button
3. Eventually receiving the nameservers to use for the CDN.

Everything in between is handled by the plugin and QC core.

This commit contains the following:
 * A button to link to a QUIC.cloud account. This will open the
QUIC.cloud login page and after logging in, it will redirect back to
wordpress.
 * A button to run CDN setup. This performs everything through DNS
detection.
 * Eventually, it will have nameservers available for the user to update
their nameservers to.

The interface could use a lot of work, but the general execution flow is
there.